Wrap os import in a try/except block
IronPython, in its infinite wisdom, does not have
the os module, so it is necessary to wrap the import (as well as the
setting of the environment variable) in a protective block.

This allows us the freedom of expanding the test suite while not
throwing an ImportError exception when installing the game.